backend/requirements.txt
fastapi
uvicorn
pydantic
googlemaps
pandas
openpyxl
python-dotenv
sse-starlette
python-multipart
